requirements.txt
aiohttp==3.8.4
cpca==0.5.5
Flask==2.2.3
Flask_Bootstrap==3.3.7.1
Flask_Login==0.6.2
flask_oidc==1.4.0
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
itsdangerous==1.1.0
joblib==1.2.0
numpy==1.19.3
openai==0.27.0
pandas==1.3.5
PyMySQL==1.0.2
python-dotenv==1.0.0
redis==4.5.1
Requests==2.29.0
SpeechRecognition==3.9.0
SQLAlchemy==1.4.35
Werkzeug==2.2.3
WTForms==3.0.1
